+++ src/index.js
@@ -1,19 +1,19 @@
-
+import {
+  BoxGeometry,
+  Mesh,
+  MeshBasicMaterial,
+  PerspectiveCamera,
+  PlaneGeometry,
+  Scene,
+  WebGLRenderer,
+  Raycaster,
+  Vector2,
+  ShaderMaterial,
+} from "three";
 
 import gsap from "gsap";
 import Scrollbar from "smooth-scrollbar";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
-import {
-  WebGLRenderer,
-  Scene,
-  PerspectiveCamera,
-  PlaneGeometry,
-  MeshBasicMaterial,
-  ShaderMaterial,
-  Mesh,
-  Raycaster,
-  Vector2,
-} from "three";
 
 const world = {};
 const os = [];
@@ -25,7 +25,7 @@ const pointer = new Vector2();
 
 init();
 function init() {
-  initScroller();
+  scrollInit();
   bindResizeEvents();
 
   world.renderer = new WebGLRenderer({
@@ -34,7 +34,7 @@ function init() {
   });
   world.renderer.setSize(canvasRect.width, canvasRect.height, false);
   world.renderer.setPixelRatio(window.devicePixelRatio);
-  world.renderer.setClearColor(0x000000, 0);
+  world.renderer.setClearColor(0x000000, 0.0);
 
   world.scene = new Scene();
 
@@ -61,28 +61,29 @@ function init() {
     // });
     const material = new ShaderMaterial({
       vertexShader: `
-        varying vec2 vUv;
+      varying vec2 vUv;
 
-        void main() {
+      void main() {
           vUv = uv;
-          gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.);
+        gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
         }
-      `,
+        `,
       fragmentShader: `
-        varying vec2 vUv;
-        uniform vec2 uMouse;
-        uniform float uHover;
+      varying vec2 vUv;
+      uniform vec2 uMouse;
+      uniform float uHover;
 
-        void main() {
-          vec2 mouse = step(uMouse, vUv);
-          gl_FragColor = vec4(mouse, uHover, 1.);
-        }
+      void main() {
+        vec2 mouse = step(uMouse, vUv);
+        gl_FragColor = vec4(mouse, uHover, 1.);
+      }
       `,
       uniforms: {
-        uMouse: { value: new Vector2(0.5,0.5) },
-        uHover: { value: 0 }
+        uMouse: { value: new Vector2(0.5, 0.5) },
+        uHover: { value: 0 },
       }
-    })
+    });
+
     const mesh = new Mesh(geometry, material);
     mesh.position.z = 0;
 
@@ -107,12 +108,10 @@ function init() {
   render();
   function render() {
     requestAnimationFrame(render);
-    // スクロール処理
-    os.forEach((o) => scroll(o));
-
-    // レイキャスティング
+    os.forEach((o) => {
+      scroll(o);
+    });
     raycast();
-
     world.renderer.render(world.scene, world.camera);
   }
 }
@@ -140,7 +139,7 @@ function resize(o, newCanvasRect) {
   mesh.position.x = x;
   mesh.position.y = y;
 
-  // 大きさの変更
+  //大きさの変更
   geometry.scale(nextRect.width / rect.width, nextRect.height / rect.height, 1);
 
   o.rect = nextRect;
@@ -152,12 +151,14 @@ function getWorldPosition(rect, canvasRect) {
   return { x, y };
 }
 
-function initScroller() {
+function scrollInit() {
   gsap.registerPlugin(ScrollTrigger);
 
   const pageContainer = document.querySelector("#page-container");
 
-  const scrollBar = Scrollbar.init(pageContainer, { delegateTo: document });
+  const scrollBar = Scrollbar.init(pageContainer, {
+    delegateTo: document,
+  });
 
   ScrollTrigger.scrollerProxy(pageContainer, {
     scrollTop(value) {
@@ -167,8 +168,13 @@ function initScroller() {
       return scrollBar.scrollTop; // getter
     },
     // getBoundingClientRect() {
-    //   return {top: 0, left: 0, width: window.innerWidth, height: window.innerHeight};
-    // }
+    //   return {
+    //     top: 0,
+    //     left: 0,
+    //     width: window.innerWidth,
+    //     height: window.innerHeight,
+    //   };
+    // },
   });
 
   scrollBar.addListener(ScrollTrigger.update);
@@ -182,21 +188,17 @@ function initScroller() {
 
 function bindResizeEvents() {
   let timerId = null;
-
   window.addEventListener("resize", () => {
     clearTimeout(timerId);
     timerId = setTimeout(() => {
       console.log("resize");
 
       const newCanvasRect = canvas.getBoundingClientRect();
-
       // canvasサイズの変更
       world.renderer.setSize(newCanvasRect.width, newCanvasRect.height, false);
-
-      // meshの位置とサイズの変更
+      // meshの位置の再計算
       os.forEach((o) => resize(o, newCanvasRect));
-
-      // cameraのProjectionMatrixの変更
+      // cameraの位置の再計算
       const cameraWidth = newCanvasRect.width;
       const cameraHeight = newCanvasRect.height;
       const near = 1500;
@@ -215,87 +217,42 @@ function bindResizeEvents() {
   });
 }
 
+function onPointerMove(event) {
+  // calculate pointer position in normalized device coordinates
+  // (-1 to +1) for both components
 
-
-function onPointerMove( event ) {
-
-	// calculate pointer position in normalized device coordinates
-	// (-1 to +1) for both components
-
-	pointer.x = ( event.clientX / window.innerWidth ) * 2 - 1;
-	pointer.y = - ( event.clientY / window.innerHeight ) * 2 + 1;
-
+  pointer.x = (event.clientX / window.innerWidth) * 2 - 1;
+  pointer.y = -(event.clientY / window.innerHeight) * 2 + 1;
 }
 
 function raycast() {
+  // update the picking ray with the camera and pointer position
+  raycaster.setFromCamera(pointer, world.camera);
 
-	// update the picking ray with the camera and pointer position
-	raycaster.setFromCamera( pointer, world.camera );
-
-	// calculate objects intersecting the picking ray
-	const intersects = raycaster.intersectObjects( world.scene.children );
+  // calculate objects intersecting the picking ray
+  const intersects = raycaster.intersectObjects(world.scene.children);
   const intersect = intersects[0];
-  
-	for ( let i = 0; i < world.scene.children.length; i ++ ) {
+
+  for (let i = 0; i < world.scene.children.length; i++) {
     const _mesh = world.scene.children[i];
 
     const uHover = _mesh.material.uniforms.uHover;
-    if(intersect?.object === _mesh) {
+    if (intersect?.object === _mesh) {
       _mesh.material.uniforms.uMouse.value = intersect.uv;
       uHover.__endValue = 1;
     } else {
       uHover.__endValue = 0;
     }
 
-    uHover.value = lerp(uHover.value, uHover.__endValue, .1);
-	}
+    uHover.value = lerp(uHover.value, uHover.__endValue, 0.1);
+  }
 }
 
-// 線形補間
+//線形補完
 function lerp(a, b, n) {
   let current = (1 - n) * a + n * b;
   if(Math.abs(b - current) < 0.001) current = b;
   return current;
 }
 
-window.addEventListener( 'pointermove', onPointerMove );
-
-/**
- * Raycasting
- * https://threejs.org/docs/#api/en/core/Raycaster
-
-const raycaster = new THREE.Raycaster();
-const pointer = new THREE.Vector2();
-
-function onPointerMove( event ) {
-
-	// calculate pointer position in normalized device coordinates
-	// (-1 to +1) for both components
-
-	pointer.x = ( event.clientX / window.innerWidth ) * 2 - 1;
-	pointer.y = - ( event.clientY / window.innerHeight ) * 2 + 1;
-
-}
-
-function render() {
-
-	// update the picking ray with the camera and pointer position
-	raycaster.setFromCamera( pointer, camera );
-
-	// calculate objects intersecting the picking ray
-	const intersects = raycaster.intersectObjects( scene.children );
-
-	for ( let i = 0; i < intersects.length; i ++ ) {
-
-		intersects[ i ].object.material.color.set( 0xff0000 );
-
-	}
-
-	renderer.render( scene, camera );
-
-}
-
-window.addEventListener( 'pointermove', onPointerMove );
-
-window.requestAnimationFrame(render);
- */
+window.addEventListener("pointermove", onPointerMove);
